third_party/WebKit/LayoutTests/battery-status/restricted-level-precision-expected.txt

Test to ensure level is reported with restricted precision.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

batteryStatusSuccess invoked
PASS battery.level is levelRounded
PASS successfullyParsed is true

TEST COMPLETE
